% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bioreg.R
\docType{data}
\name{bioreg_list}
\alias{bioreg_list}
\title{Bioreg Data Files}
\format{
A list with each element containing:
\describe{
  \item{filename}{The name to save the downloaded file as.}
  \item{link}{The URL to download the file.}
  \item{description}{A short description of the file.}
}
}
\usage{
bioreg_list
}
\description{
A list of data files including download links, filenames, and descriptions.
}
\examples{
\dontrun{
bioreg_list
}
}
\keyword{datasets}
