% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_character_vector}
\alias{check_character_vector}
\title{Check Whether an Argument Is a Character Vector}
\usage{
check_character_vector(x, name)
}
\arguments{
\item{x}{Argument value.}

\item{name}{Argument name used in error messages.}
}
\description{
Check Whether an Argument Is a Character Vector
}
\keyword{internal}
