% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_logical}
\alias{check_logical}
\title{Check Whether an Argument Is a Logical Scalar}
\usage{
check_logical(x, name)
}
\arguments{
\item{x}{Argument value.}

\item{name}{Argument name used in error messages.}
}
\description{
Check Whether an Argument Is a Logical Scalar
}
\keyword{internal}
