% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_null_na}
\alias{check_null_na}
\title{Check Whether an Argument Is Missing}
\usage{
check_null_na(x, name)
}
\arguments{
\item{x}{Argument value.}

\item{name}{Argument name used in error messages.}
}
\description{
Check Whether an Argument Is Missing
}
\keyword{internal}
