% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_numeric}
\alias{check_numeric}
\title{Check Whether an Argument Is a Numeric Scalar}
\usage{
check_numeric(x, name)
}
\arguments{
\item{x}{Argument value.}

\item{name}{Argument name used in error messages.}
}
\description{
Check Whether an Argument Is a Numeric Scalar
}
\keyword{internal}
