% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_numeric_range}
\alias{check_numeric_range}
\title{Check Whether an Argument Is a Numeric Vector of Fixed Length}
\usage{
check_numeric_range(x, name, len)
}
\arguments{
\item{x}{Argument value.}

\item{name}{Argument name used in error messages.}

\item{len}{Expected length of the numeric vector.}
}
\description{
Check Whether an Argument Is a Numeric Vector of Fixed Length
}
\keyword{internal}
