% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bioreg.R
\name{get_bioreg}
\alias{get_bioreg}
\title{Download Bioreg Data Files}
\usage{
get_bioreg(bioreg_name = "all", save_dir = NULL)
}
\arguments{
\item{bioreg_name}{"all" or the filename of the desired bioregion to be
downloaded. See `bioreg_list` for available options. This contains a list
of data file information with elements containing `link`, `filename`,
and `description`.}

\item{save_dir}{The directory to save the downloaded files. Defaults to
"inst/extdata/downloads" within the package structure.}
}
\value{
NULL. Downloads the files to the specified directory.
}
\description{
Downloads data files from a provided list of links and saves them
to a specified directory.
}
\examples{
\dontrun{
# download all bioregions available in bioreg_list
get_bioreg()
}
}
