% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{getGBIF}
\alias{getGBIF}
\title{Constructor for \code{getGBIF} Objects}
\usage{
getGBIF(df)
}
\arguments{
\item{df}{A data frame containing GBIF occurrence records.}
}
\value{
An object of class \code{getGBIF}.
}
\description{
Wrap a data frame in the \code{getGBIF} class used by downstream functions
in the package.
}
\keyword{internal}
