% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{getRange-class}
\alias{getRange-class}
\alias{getRange}
\title{Reference Class for \code{get_range()} Results}
\value{
A generator object of reference class \code{"getRange"}, used to
instantiate objects with two fields: \code{init.args} (the original
arguments used to build the range map) and \code{rangeOutput} (the
resulting spatial output).
}
\description{
Stores the original arguments used to build a range map and the resulting
spatial output.
}

