% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{log_step}
\alias{log_step}
\title{Append a Filtering Step to a Log Table}
\usage{
log_step(log, step_name, before, after)
}
\arguments{
\item{log}{Existing log data frame.}

\item{step_name}{Name of the filtering step.}

\item{before}{Number of records before filtering.}

\item{after}{Number of records after filtering.}
}
\value{
An updated log data frame.
}
\description{
Append a Filtering Step to a Log Table
}
\keyword{internal}
