% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbif_file_workflows.R
\name{plot.gbifPackedSpatRaster}
\alias{plot.gbifPackedSpatRaster}
\title{Plot a Packed Raster Range Saved by \code{species_csvs_to_ranges()}}
\usage{
\method{plot}{gbifPackedSpatRaster}(x, ...)
}
\arguments{
\item{x}{Object of class \code{gbifPackedSpatRaster}.}

\item{...}{Additional arguments passed to \code{terra::plot()}.}
}
\description{
Plot a Packed Raster Range Saved by \code{species_csvs_to_ranges()}
}
\keyword{internal}
