% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbif_file_workflows.R
\name{plot.gbifPackedSpatVector}
\alias{plot.gbifPackedSpatVector}
\title{Plot a Packed Vector Range Saved by \code{species_csvs_to_ranges()}}
\usage{
\method{plot}{gbifPackedSpatVector}(x, ...)
}
\arguments{
\item{x}{Object of class \code{gbifPackedSpatVector}.}

\item{...}{Additional arguments passed to \code{terra::plot()}.}
}
\description{
Plot a Packed Vector Range Saved by \code{species_csvs_to_ranges()}
}
\keyword{internal}
