% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bioreg.R
\name{read_bioreg}
\alias{read_bioreg}
\title{Load Bioreg Data Files}
\usage{
read_bioreg(bioreg_name = "eco_terra", save_dir = NULL)
}
\arguments{
\item{bioreg_name}{filename of the desired bioregion to be loaded. See `bioreg_list` for available options.}

\item{save_dir}{The directory to save the downloaded files. Defaults to "inst/extdata/downloads" within the package structure.}
}
\value{
A terra::vect object representing the shapefile.
}
\description{
Loads a shapefile based on the provided bioregion name and the bioreg_list.
}
\examples{
\dontrun{
shp_eco_terra <- read_bioreg("eco_terra")
plot(shp_eco_terra)
}
}
