% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsl_taXnames.R
\name{wsl.taXnames}
\alias{wsl.taXnames}
\title{Retrieve from GBIF all scientific names of a specific Taxa}
\usage{
wsl.taXnames(sp_name = NULL, conf_match = 90, all = FALSE)
}
\arguments{
\item{sp_name}{Character. Species name from which the user wants to retrieve all existing GBIF names}

\item{conf_match}{Numeric. From 0 to 100. Determine the confidence
threshold of match of 'sp_name' with the GBIF backbone taxonomy. Default is 90.}

\item{all}{Logical. Default is FALSE. Should all species names be retrieved or only
the accepted name and its synonyms?}
}
\value{
A data.frame with two columns: (1) Names and (2) Backbone Taxonomy Status
}
\description{
Allows to extract from the gbif backbone taxonomy all names from an input species name
(accepted, synonyms, children, related...).
}
\examples{
wsl.taXnames("Cypripedium calceolus",all=FALSE)
wsl.taXnames("Cypripedium calceolus",all=TRUE)

}
\author{
Yohann Chauvier
}
