% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsl_obs_filter.R
\name{wsl_obs_filter}
\alias{wsl_obs_filter}
\title{Filter a set of GBIF observations according to a defined grain}
\usage{
wsl_obs_filter(wsl.gbif, grid)
}
\arguments{
\item{wsl.gbif}{one wsl.gbif output including one or several species}

\item{grid}{Object of class 'SpatRaster', 'RasterLayer', 'RasterBrick' or
'RasterStack' of desired resolution and extent.}
}
\value{
a data frame with two columns named 'x' and 'y' comprising
the new set of observations filtered at grid resolution.
}
\description{
Filter a set of GBIF observations through a chosen raster grid that defines
the resolution of the spatial analysis. Depending on the desired resolution
of the analysis, one might want to keep only one GBIF observation per grid
cell to avoid e.g., modelling and sampling bias issues.
}
\examples{

# Load the European Alps extent and a raster of a random resolution
data(geo_dat)
data(exrst)

# Downloading in the European Alps the observations of Arctostaphylos alpinus
obs.arcto = wsl_gbif("Arctostaphylos alpinus",geo=shp.lonlat)
plot(shp.lonlat)
points(obs.arcto[,c("decimalLongitude","decimalLatitude")],pch=20,col="#238b4550",cex=1)

}
