% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsl_taxonomy.R
\name{wsl_taxonomy}
\alias{wsl_taxonomy}
\title{Retrieve from GBIF all scientific names of a specific Taxa}
\usage{
wsl_taxonomy(sp_name = NULL, conf_match = 90, all = FALSE)
}
\arguments{
\item{sp_name}{Character. Species name from which the user wants to retrieve all existing GBIF names}

\item{conf_match}{Numeric. From 0 to 100. Determine the confidence
threshold of match of 'sp_name' with the GBIF backbone taxonomy. Default is 90.}

\item{all}{Logical. Default is FALSE. Should all species names be retrieved or only
the accepted name and its synonyms?}
}
\value{
A data.frame with two columns: (1) Names and (2) Backbone Taxonomy Status
}
\description{
Generates, based on a given species name, a list of all its scientific names
(accepted, synonyms) found in the GBIF backbone taxonomy to download the data.
Children and related doubtful names not use to download the data may also be extracted.
}
\examples{
wsl_taxonomy("Cypripedium calceolus",all=FALSE)
wsl_taxonomy("Cypripedium calceolus",all=TRUE)

}
\references{
Chamberlain, S., Oldoni, D., & Waller, J. (2022). rgbif: interface to the global biodiversity
information facility API. 10.5281/zenodo.6023735
}
\seealso{
The rgbif package for additional and more general approaches on how to retrieve
scientific names from the GBIF backbone taxonomy.
}
