% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ecoreg.R
\name{check_and_get_ecoreg}
\alias{check_and_get_ecoreg}
\title{Check for a Local Ecoregion Layer and Download It if Needed}
\usage{
check_and_get_ecoreg(ecoreg_name = "eco_terra", save_dir = NULL)
}
\arguments{
\item{ecoreg_name}{File name of the ecoregion dataset to check. See
\code{ecoreg_list} for valid values.}

\item{save_dir}{Directory where downloaded files should be stored.}
}
\value{
\code{NULL}. The function downloads data only when required.
}
\description{
Check whether a target ecoregion directory exists and contains at least one
shapefile. If not, download the dataset with \code{get_ecoreg()}.
}
\examples{
\dontrun{
check_and_get_ecoreg("eco_terra")
}
}
