% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ecoreg.R
\name{check_and_get_ecoreg}
\alias{check_and_get_ecoreg}
\title{Check and Download Ecoreg Data Files}
\usage{
check_and_get_ecoreg(ecoreg_name = "eco_terra", save_dir = NULL)
}
\arguments{
\item{ecoreg_name}{Filename of the desired ecoregion to be checked and
downloaded if necessary. See \code{ecoreg_list} for available options.}

\item{save_dir}{The directory to save the downloaded files. Defaults to
"inst/extdata/downloads" within the package structure.}
}
\value{
\code{NULL}. Checks and downloads the files to the specified
directory if necessary.
}
\description{
Checks if a directory exists and contains at least one .shp file.
If not, calls get_ecoreg to download the data.
}
\examples{
\dontrun{
check_and_get_ecoreg("eco_terra")
}
}
