% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_character_vector}
\alias{check_character_vector}
\title{Check if argument is a non-empty character vector (no NA)}
\usage{
check_character_vector(x, name)
}
\arguments{
\item{x}{Argument value}

\item{name}{Argument name for error reporting}
}
\description{
Check if argument is a non-empty character vector (no NA)
}
\keyword{internal}
