% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_logical}
\alias{check_logical}
\title{Check if argument is a non-missing logical scalar}
\usage{
check_logical(x, name)
}
\arguments{
\item{x}{Argument value}

\item{name}{Argument name for error reporting}
}
\description{
Check if argument is a non-missing logical scalar
}
\keyword{internal}
