% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_null_na}
\alias{check_null_na}
\title{Check if an argument is NULL or NA}
\usage{
check_null_na(x, name)
}
\arguments{
\item{x}{Argument value}

\item{name}{Argument name for error reporting}
}
\description{
Check if an argument is NULL or NA
}
\keyword{internal}
