% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_numeric}
\alias{check_numeric}
\title{Check if argument is a non-missing numeric scalar}
\usage{
check_numeric(x, name)
}
\arguments{
\item{x}{Argument value}

\item{name}{Argument name for error reporting}
}
\description{
Check if argument is a non-missing numeric scalar
}
\keyword{internal}
