% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_numeric_range}
\alias{check_numeric_range}
\title{Check if argument is a numeric vector of exact length}
\usage{
check_numeric_range(x, name, len)
}
\arguments{
\item{x}{Argument value}

\item{name}{Argument name for error reporting}

\item{len}{Expected length of numeric vector}
}
\description{
Check if argument is a numeric vector of exact length
}
\keyword{internal}
