% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_function.R
\name{conv_function}
\alias{conv_function}
\title{Create polygon objects in different bioregions.}
\usage{
conv_function(sp_coord, bwp, bipl, bwpo, temp_dir, g = NULL)
}
\arguments{
\item{sp_coord}{Spatial coordinates (type "matrix" or "data.frame).}

\item{bwp}{Buffer width parameter (type numeric).}

\item{bipl}{Number of observation points (type numeric).}

\item{bwpo}{Buffer width parameter for the convex hull (type numeric).}

\item{temp_dir}{Temporary directory name (type character).}

\item{g}{Optional parameter.}
}
\description{
Not to be called directly by the user.
}
\keyword{internal}
