% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_function.R
\name{conv_function}
\alias{conv_function}
\title{Create polygon objects in different ecoregions.}
\usage{
conv_function(sp_coord, bwp, bipl, bwpo, temp_dir, g = NULL)
}
\arguments{
\item{sp_coord}{Object of class matrix of data.frame. Spatial coordinates.}

\item{bwp}{Numeric. Buffer width parameter.}

\item{bipl}{Numeric. Number of observation points.}

\item{bwpo}{Numeric. Buffer width parameter for the convex hull.}

\item{temp_dir}{Character. Temporary directory name.}

\item{g}{Optional parameter.}
}
\description{
Not to be called directly by the user.
}
\keyword{internal}
