% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{cscl}
\alias{cscl}
\title{Original legend bar (from Dr. Philipp Brun)}
\usage{
cscl(
  colors,
  crds,
  horiz = FALSE,
  zrng = c(0, 100),
  at = 10 * 0:10,
  labs = NA,
  tickle = 0.2,
  title = 1,
  lablag = 1,
  titlag = 2,
  box = TRUE,
  breaks,
  cx = 0.8,
  tria = "n"
)
}
\arguments{
\item{colors}{Parameter 1}

\item{crds}{Parameter 2}

\item{horiz}{Parameter 3}

\item{zrng}{Parameter 5}

\item{at}{Parameter 6}

\item{labs}{Parameter 7}

\item{tickle}{Parameter 8}

\item{title}{Parameter 9}

\item{lablag}{Parameter 10}

\item{titlag}{Parameter 11}

\item{box}{Parameter 12}

\item{breaks}{Parameter 13}

\item{cx}{Parameter 14}

\item{tria}{Parameter 15}
}
\description{
Original legend bar (from Dr. Philipp Brun)
}
