% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ecoreg.R
\docType{data}
\name{ecoreg_list}
\alias{ecoreg_list}
\title{Metadata for Downloadable Ecoregion Layers}
\format{
A list with each element containing:
\describe{
  \item{filename}{The name to save the downloaded file as.}
  \item{link}{The URL to download the file.}
  \item{description}{A short description of the file.}
}
}
\usage{
ecoreg_list
}
\description{
A list describing the ecoregion layers that can be downloaded with
\code{get_ecoreg()}.
}
\examples{
\dontrun{
ecoreg_list
}
}
\keyword{datasets}
