% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fig_label}
\alias{fig_label}
\title{Custom figure label plot}
\usage{
fig_label(
  text,
  region = "figure",
  pos = "topleft",
  cex = NULL,
  margin = 0.02,
  ...
)
}
\arguments{
\item{text}{Parameter 1}

\item{region}{Parameter 2}

\item{pos}{Parameter 3}

\item{cex}{Parameter 5}

\item{margin}{Parameter 6}

\item{...}{Parameter 7}
}
\description{
Custom figure label plot
}
\keyword{internal}
