% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_doi.R
\name{get_doi}
\alias{get_doi}
\title{Get a custom DOI for a GBIF filtered dataset}
\usage{
get_doi(
  gbifs = NULL,
  title = NULL,
  description = NULL,
  source_url = "https://example.com/",
  user = "",
  pwd = "",
  ...
)
}
\arguments{
\item{gbifs}{Object of class getGBIF or a list of several.}

\item{title}{Character. Chosen title for the GBIF dataset.}

\item{description}{Character. Description of the GBIF dataset.}

\item{source_url}{Character. Link to where the dataset is stored.}

\item{user}{Character. User's GBIF username.}

\item{pwd}{Character. User's GBIF password.}

\item{...}{Additonnal parameters for \code{derived_dataset()} called
from the \code{rgbif} R package.}
}
\value{
One citable GBIF DOI and associated user's information.
}
\description{
A small user friendly wrapper of \code{derived_dataset()}
(\code{rgbif} R package), compatible with one or several
\code{get_gbif()} outputs.
}
\details{
See \code{derived_dataset()} from the \code{rgbif} R package.
}
\examples{
# Download worldwide the observations of Panthera tigris
# and Ailuropoda melanoleuca
obspt <- get_gbif("Panthera tigris")
obsam <- get_gbif("Ailuropoda melanoleuca")

\dontrun{
# Retrieve DOI for only one get_gbif() output
get_doi(
       gbifs = obspt,
       title = "GBIF_test1",
       description = "A small example 1",
       source_url = "https://example.com/",
       user = "",
       pwd = ""
) # Use your own GBIF credentials here

# Retrieve DOIs for several get_gbif() outputs
get_doi(
       gbifd = list(obspt,obsam),
       title = "GBIF_test2",
       description = "A small example 2",
       source_url = "https://example.com/",
       user = "",
       pwd = ""
) # Use your own GBIF credentials here

}
}
\references{
Chamberlain, S., Oldoni, D., & Waller, J. (2022). rgbif: interface to the
global biodiversity information facility API. 10.5281/zenodo.6023735
}
\seealso{
The \code{rgbif package} for additional and more general approaches
to get GBIF DOI.
}
