% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_doi.R
\name{get_doi}
\alias{get_doi}
\title{Get a custom DOI for a GBIF filtered dataset}
\usage{
get_doi(
  gbifs = NULL,
  title = NULL,
  description = NULL,
  source_url = "https://example.com/",
  user = "",
  pwd = "",
  ...
)
}
\arguments{
\item{gbifs}{data.frame or list. One get_gbif() output or a list of several.}

\item{title}{Title for your derived dataset.}

\item{description}{Description of the dataset.}

\item{source_url}{Link to where the dataset is stored.}

\item{user}{Your GBIF username.}

\item{pwd}{Your GBIF password.}

\item{...}{Additonnal parameters for derived_dataset() in rgbif.}
}
\value{
One citable DOI and its information.
}
\description{
A small user friendly wrapper of the derived_dataset() function of the
rgbif R package, compatible with one or several get_gbif() outputs.
}
\details{
see derived_dataset() function from the rgbif R package.
}
\examples{
# Download worldwide the observations of Panthera tigris and Ailuropoda melanoleuca
obs.pt <- get_gbif("Panthera tigris")
obs.am <- get_gbif("Ailuropoda melanoleuca")
\dontrun{
# Retrieve DOI for only one get_gbif() output
get_doi(obs.pt,title="GBIF_test1",description="A small example 1",
   source_url="https://example.com/",user="",pwd="") # Use your own GBIF credentials here

# Retrieve DOIs for several get_gbif() outputs
get_doi(list(obs.pt,obs.am),title="GBIF_test2",description="A small example 2",
   source_url="https://example.com/",user="",pwd="") # Use your own GBIF credentials here
}
}
\references{
Chamberlain, S., Oldoni, D., & Waller, J. (2022). rgbif: interface to the global biodiversity
information facility API. 10.5281/zenodo.6023735
}
\seealso{
The rgbif package for additional and more general approaches to get GBIF DOI.
}
