% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ecoreg.R
\name{get_ecoreg}
\alias{get_ecoreg}
\title{Download Ecoreg Data Files}
\usage{
get_ecoreg(ecoreg_name = "all", save_dir = NULL)
}
\arguments{
\item{ecoreg_name}{Character. "all" or the filename of the desired ecoregion
to be downloaded. See \code{ecoreg_list} for available options. This
contains a list of data file information with elements containing `link`,
`filename`, and `description`.}

\item{save_dir}{The directory to save the downloaded files. Defaults to
"inst/extdata/downloads" within the package structure.}
}
\value{
\code{NULL}. Downloads the files to the specified directory.
}
\description{
Downloads data files from a provided list of links and saves them
to a specified directory.
}
\examples{
\dontrun{
# download all ecoregions available in ecoreg_list
get_ecoreg()
}
}
