% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gbif_count.R
\name{get_gbif_count}
\alias{get_gbif_count}
\title{Retrieve GBIF observations count to select study case}
\usage{
get_gbif_count(
  sp_name = NULL,
  search = TRUE,
  rank = NULL,
  phylum = NULL,
  class = NULL,
  order = NULL,
  family = NULL,
  conf_match = 80,
  geo = NULL,
  has_xy = TRUE,
  spatial_issue = FALSE
)
}
\arguments{
\item{sp_name}{Character. Species name may be scientific or addressing
Genus and species only. Fuzzy matches are allowed but see \code{search}
parameter below.}

\item{search}{Logical. If \code{TRUE} (default), the function will strictly
look for the most relevant result across the names given by \code{rgbif}
(only species, subspecies and variety allowed here), and give an error if
name matching is impeded by synonym duplicates. If \code{FALSE}, the
function will simply pick the first most relevant name (taxa rank higher
than species are allowed here), with fuzzy search (~approximative name
matching) allowed, and \code{rank},\code{phylum}, \code{class},
\code{order} and \code{family} parameters optionally used, i.e., only
if no convincing match is found. \code{FALSE} is particularly useful if
the given species name already includes the author.}

\item{rank}{Character. "SPECIES", "SUBSPECIES" or "VARIETY". If \code{NULL}
(default), the order of priority is (1) species, (2) subspecies and (3)
variety unless "subsp." or "var." is found in the \code{sp_name} parameter.}

\item{phylum}{Character (optional). What is the species' Phylum? Adds a
criteria to deal with alternative name matches and select the right species.
Particularly useful to avoid hemihomonyms (a scientific name used for
different taxa). Available options are the GBIF Phylums
(\href{https://www.gbif.org/species/search}{listed per Kingdom/Phylum}).
If \code{search = FALSE}, only used if no direct match is found.}

\item{class}{Character (optional). What is the species' Class? Same as above
but at the finer class level. Available options are the GBIF Classes
(same url). If \code{search = FALSE}, only used if no direct match is found.}

\item{order}{Character (optional). What is the species' Order? Same as above
but at the finer order level. Available options are the GBIF Orders (same
url). If \code{search = FALSE}, only used if no direct match is found.}

\item{family}{Character (optional). What is the species' Family? Same as
above but at the finer family level. Available options are the GBIF Families
(same url). If \code{search = FALSE}, used only if no direct match is found.}

\item{conf_match}{Numeric from 0 to 100. Determine the confidence threshold
of match between \code{sp_name} and the GBIF backbone taxonomy.
Default is 90.}

\item{geo}{Object of class \code{Extent}, \code{SpatExtent},
\code{SpatialPolygon}, \code{SpatialPolygonDataframe}, \code{SpatVector}
or \code{sf} (WGS84) to define the study's area extent. Default is
\code{NULL}, i.e., the whole globe.}

\item{has_xy}{Logical. If \code{TRUE}, only records with coordinates are
accounted ofr (default). If \code{FALSE}, only records without coordinates.
If \code{NULL}, all records.}

\item{spatial_issue}{Logical. If \code{FALSE}, only records without
spatial issues are accounted for (default). If \code{TRUE}, only records with
spatial issues. If \code{NULL}, all records.}
}
\value{
A simple print of the number of observations found.
}
\description{
Follows the same code implementation as \code{get_gbif} for retrieving
the number of occurrences per species without downloading the data.
}
\details{
Implements the same search result when
(\href{https://www.gbif.org}{GBIF}) is employed, i.e., based on the
input taxa name, all species records count related to its accepted name
and synonyms are extracted.
}
\examples{
# Get number of observations with default (only XY with no spatial issues):
obs.pt <- get_gbif_count(
	sp_name = "Ailuropoda melanoleuca",
	has_xy = TRUE,
	spatial_issue = FALSE,
	geo = NULL
)

# Get total number of observations (everything):
obs.pt <- get_gbif_count(
	sp_name = "Ailuropoda melanoleuca",
	has_xy = NULL,
	spatial_issue = NULL,
	geo = NULL
)

# Example of setting 'geo' (has_xy = TRUE is forced)
obs.pt <- get_gbif_count(
	sp_name = "Ailuropoda melanoleuca",
	has_xy = NULL,
	spatial_issue = NULL,
	geo = terra::ext()
)

# Example of fuzzy match when search is set to FALSE
obs.pt <- get_gbif_count(
	sp_name = "Ailuropoda melanolca",
	search = FALSE
)
}
\references{
Chamberlain, S., Oldoni, D., & Waller, J. (2022). rgbif: interface to the
global biodiversity information facility API. 10.5281/zenodo.6023735
}
