% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_range.R
\name{get_range}
\alias{get_range}
\title{Create a species range map based on a get_gbif() output}
\usage{
get_range(
  species_name = NULL,
  occ_coord = NULL,
  Bioreg = NULL,
  Bioreg_name = NULL,
  degrees_outlier = 10,
  clustered_points_outlier = 3,
  buffer_width_point = 0.5,
  buffer_increment_point_line = 0.5,
  buffer_width_polygon = 0.1,
  dir_temp = paste0("temp", sample(1:99999999, 1))
)
}
\arguments{
\item{species_name}{character string of the species name. E.g. "Anemone nemorosa".}

\item{occ_coord}{a SpatialPoints object.}

\item{Bioreg}{shapefile containg different bioregions (convex hulls will be classified
on a bioreg basis).}

\item{Bioreg_name}{how is the slot containing the bioregion names called?}

\item{degrees_outlier}{distance threshold (degrees) for outlier classification. If the
nearest minimal distance to the next point is larger than this threshold, it will be
considered as an outlier.}

\item{clustered_points_outlier}{maximum number of points which are closer to each other
than the degrees_outlier, but should still be considered as outliers.}

\item{buffer_width_point}{buffer (in degrees) which will be applied around single observations.}

\item{buffer_increment_point_line}{how much should the buffer be increased for each point on a line.}

\item{buffer_width_polygon}{buffer (in degrees) which will be applied around distribution polygons
(for each bioregion).}

\item{dir_temp}{where should the temporary text file for the convex hull be saved?
(text file will be deleted again).}
}
\value{
a shapefile.
}
\description{
Estimates species ranges based on occurrence data (GBIF or not) and bioregions.
It first deletes outliers from the observation dataset and then creates a polygon
(convex hull) with a user specified buffer around all the observations of one bioregion.
If there there is only one observation in a bioregion, a buffer around this point
will be created. If all points in a bioregion are on a line, the function will also
create a buffer around these points, however, the buffer size increases with the number
of points in the line.
}
\details{
...
}
\references{
Lyu, L., Leugger, F., Hagen, O., Fopp, F., Boschman, L. M., Strijk, J. S., ... &
Pellissier, L. (2022). An integrated high‐resolution mapping shows congruent biodiversity
patterns of Fagales and Pinales. New Phytologist, 235(2), 759-772 10.1111/nph.18158

Hijmans, Robert J. "terra: Spatial Data Analysis. R Package Version 1.6-7." (2022). Terra - CRAN
}
\seealso{
...
}
