% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_range.R
\name{get_range}
\alias{get_range}
\title{Create a species range map based on a get_gbif() output}
\usage{
get_range(
  sp_name = NULL,
  occ_coord = NULL,
  Bioreg = eco.earth,
  Bioreg_name = "ECO_NAME",
  degrees_outlier = 3,
  clustered_points_outlier = 2,
  buffer_width_point = 4,
  buffer_increment_point_line = 0.5,
  buffer_width_polygon = 4,
  dir_temp = paste0("temp", sample(1:99999999, 1)),
  raster = TRUE,
  res = 10
)
}
\arguments{
\item{sp_name}{Character. Species name e.g., 'Anemone nemorosa'.}

\item{occ_coord}{a get_gbif() output or a data.frame containing two columns named
"decimalLongitude" and "decimalLatitude".}

\item{Bioreg}{'SpatialPolygonsDataFrame', 'SpatVector' or 'sf' object containg different
ecoregions (convex hulls will be classified on a bioreg basis) and of CRS WGS84. Note
that this parameter may be fed with an external, generated (function make_ecoregion) or
in-house ecoregion shapefile. Three in-house shapefiles are already included in the library:
'eco.earh' (for terrestrial species; Nature conservancy version adapted from Olson & al. 2001),
'eco.marine' (for marine species; Spalding & al. 2007, 2012) and 'eco.fresh' (for freshwater
species; Abell & al. 2008). For marine species, eco.earth may also be used if the user wants
to represent the terrestrial range of species that also partially settle on mainland. For
fresh water species, same may be done if the user considers that terrestrial ecoregions
should be more representtaive of the species ecology.}

\item{Bioreg_name}{Character. How is the slot containing the ecoregion names called?
Default is the very detailed level of 'eco.earth' (aka 'ECO_NAME'). Note that 'EcoRegion'
must always be used when using a make_ecoregion() output.}

\item{degrees_outlier}{Numeric. Distance threshold (degrees) for outlier classification.
If the nearest minimal distance to the next point is larger than this threshold, it will be
considered as an outlier.}

\item{clustered_points_outlier}{Numeric. Maximum number of points which are closer to each other
than the degrees_outlier, but should still be considered as outliers.}

\item{buffer_width_point}{Numeric. Buffer (in degrees) which will be applied around single
observations.}

\item{buffer_increment_point_line}{Numeric. How much should the buffer be increased for each point
on a line.}

\item{buffer_width_polygon}{Numeric. Buffer (in degrees) which will be applied around distribution
polygons (for each ecoregion).}

\item{dir_temp}{Character. Where should the temporary text file for the convex hull be saved?
(text file will be deleted again).}

\item{raster}{Logical. Should the output be a unified raster? Default is TRUE}

\item{res}{Numeric. If raster = TRUE, which resolution? Final resolution in ° = 1°/res
e.g.,  = 0.1° if res = 10. Default is 10.}
}
\value{
A 'SpatVector' or 'SpatRaster'.
}
\description{
Estimates species ranges based on occurrence data (GBIF or not) and ecoregions
(may be an external, in-house or make_ecoregion() input). It first deletes outliers
from the observation dataset and then creates a polygon (convex hull) with a user
specified buffer around all the observations of one ecoregion. If there is only
one observation in an ecoregion, a buffer around this point will be created. If
all points in an ecoregion are on a line, the function will also create a buffer
around these points, however, the buffer size increases with the number of points
in the line. Finally, also note that in case of too many records, get_range can be
used with a sub-sample of species observations to ensure a faster polygon process
and/or to overcome potential RAM crash of the function.
}
\details{
Ecoregions cover relatively large areas of land or water, and contain characteristic,
geographically distinct assemblages of natural communities sharing a large majority of species,
dynamics, and environmental conditions. The biodiversity of flora, fauna and ecosystems that
characterise an ecoregion tends to be distinct from that of other ecoregions
(https://en.wikipedia.org/wiki/Ecoregion).
}
\examples{
# Load available ecoregions
data(ecoregions)

# First download the worldwide observations of Panthera tigris and convert to SpatialPoints
obs.pt = get_gbif("Panthera tigris",basis=c("OBSERVATION","HUMAN_OBSERVATION","MACHINE_OBSERVATION"))

# Plot
plot(eco.earth)
plot(sp.shp,pch=20,col="#238b4550",cex=4,add=TRUE)

# Generate the distributional range map of Panthera tigris for the finest terrestrial ecoregions
range.tiger = get_range("Panthera tigris",sp.shp,eco.earth,"ECO_NAME")

# Plotting
plot(eco.earth)
plot(range.tiger,col="#238b45",add=TRUE)

}
\references{
Oskar Hagen, Lisa Vaterlaus, Camille Albouy, Andrew Brown, Flurin Leugger, Renske E. Onstein,
Charles Novaes de Santana, Christopher R. Scotese, Loïc Pellissier. (2019) Mountain building,
climate cooling and the richness of cold-adapted plants in the Northern Hemisphere. Journal of
Biogeography. doi: 10.1111/jbi.13653

Lyu, L., Leugger, F., Hagen, O., Fopp, F., Boschman, L. M., Strijk, J. S., ... &
Pellissier, L. (2022). An integrated high resolution mapping shows congruent biodiversity
patterns of Fagales and Pinales. New Phytologist, 235(2), 759-772 10.1111/nph.18158

Olson, D. M., Dinerstein, E., Wikramanayake, E. D., Burgess, N. D., Powell, G. V. N.,
Underwood, E. C., D'Amico, J. A., Itoua, I., Strand, H. E., Morrison, J. C., Loucks, C. J.,
Allnutt, T. F., Ricketts, T. H., Kura, Y., Lamoreux, J. F., Wettengel, W. W., Hedao, P., Kassem,
K. R. 2001. Terrestrial ecoregions of the world: a new map of life on Earth.
Bioscience 51(11):933-938. doi: 10.1641/0006-3568(2001)051

The Nature Conservancy (2009). Global Ecoregions, Major Habitat Types, Biogeographical Realms
and The Nature Conservancy Terrestrial Assessment Units. GIS layers developed by The Nature
Conservancy with multiple partners, combined from Olson et al. (2001), Bailey 1995 and Wiken 1986.
Cambridge (UK): The Nature Conservancy.

Mark D. Spalding, Helen E. Fox, Gerald R. Allen, Nick Davidson, Zach A. Ferdaña, Max
Finlayson, Benjamin S. Halpern, Miguel A. Jorge, Al Lombana, Sara A. Lourie, Kirsten D.
Martin, Edmund McManus, Jennifer Molnar, Cheri A. Recchia, James Robertson, Marine
Ecoregions of the World: A Bioregionalization of Coastal and Shelf Areas, BioScience,
Volume 57, Issue 7, July 2007, Pages 573–583. doi: 10.1641/B570707

Spalding, M. D., Agostini, V. N., Rice, J., & Grant, S. M. (2012). Pelagic provinces of the world:
a biogeographic classification of the world’s surface pelagic waters. Ocean & Coastal Management,
60, 19-30. doi: 10.1016/j.ocecoaman.2011.12.016

The Nature Conservancy (2012). Marine Ecoregions and Pelagic Provinces of the World. GIS layers
developed by The Nature Conservancy with multiple partners, combined from Spalding et al. (2007)
and Spalding et al. (2012). Cambridge (UK): The Nature Conservancy.

Robin Abell, Michele L. Thieme, Carmen Revenga, Mark Bryer, Maurice Kottelat, Nina Bogutskaya,
Brian Coad, Nick Mandrak, Salvador Contreras Balderas, William Bussing, Melanie L. J. Stiassny,
Paul Skelton, Gerald R. Allen, Peter Unmack, Alexander Naseka, Rebecca Ng, Nikolai Sindorf,
James Robertson, Eric Armijo, Jonathan V. Higgins, Thomas J. Heibel, Eric Wikramanayake,
David Olson, Hugo L. López, Roberto E. Reis, John G. Lundberg, Mark H. Sabaj Pérez,
Paulo Petry, Freshwater Ecoregions of the World: A New Map of Biogeographic Units for
Freshwater Biodiversity Conservation, BioScience, Volume 58, Issue 5, May 2008,
Pages 403–414. doi: 10.1641/B580507

Hijmans, Robert J. "terra: Spatial Data Analysis. R Package Version 1.6-7." (2022). Terra - CRAN
}
\seealso{
For more information on the original code and methods, check Hagen, Oskar et al. (2019), Data
from: Mountain building, climate cooling and the richness of cold-adapted plants in the northern
hemisphere, Dryad, Dataset, https://doi.org/10.5061/dryad.0ff6b04.
}
