% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{log_step}
\alias{log_step}
\title{Summary log helper for get_gbif filtering}
\usage{
log_step(log, step_name, before, after)
}
\arguments{
\item{log}{Argument value}

\item{step_name}{Filtering step name}

\item{before}{Pre-filtering number of observations}

\item{after}{Post-filtering number of observations}
}
\description{
Summary log helper for get_gbif filtering
}
\keyword{internal}
