% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_blocks.R
\name{make_blocks}
\alias{make_blocks}
\title{Split Data into Approximately Balanced Folds}
\usage{
make_blocks(
  nfolds = 5,
  df = data.frame(),
  nblocks = nfolds * 2,
  npoints = NA,
  pres = numeric()
)
}
\arguments{
\item{nfolds}{Numeric. Number of folds to create.}

\item{df}{Optional \code{data.frame}. Columns define the structure used
to build folds. When omitted, random folds are created from
\code{npoints}.}

\item{nblocks}{Numeric. Number of initial clusters used to build the folds.
Must be at least \code{nfolds}.}

\item{npoints}{Optional numeric. Number of observations to split when \code{df} is
not supplied.}

\item{pres}{Optional binary vector. Used to restrict CLARA clustering to a
subset of rows and assign the remainder by nearest neighbours.}
}
\value{
An integer vector of length \code{nrow(df)} or \code{npoints},
giving the fold assignment for each observation.
}
\description{
Create a fold-assignment vector for random or spatially structured
cross-validation.
}
\details{
If \code{df} has one column, folds are based on quantile bins. If
\code{df} has two or more columns, folds are based on CLARA clustering.
Remaining clusters are assigned to folds with an optimization step that
tries to balance fold sizes as evenly as possible.
}
\examples{
\donttest{
# Downloading worldwide all the observations of great panda 
obs.pd <- get_gbif(sp_name = "Ailuropoda melanoleuca")

# Create a vector of folds (n = 5) spatially blocked (n = 10)
block.pd <- make_blocks(
    nfolds = 5,
    df = obs.pd[, c("decimalLatitude","decimalLongitude")],
    nblocks = 5
)

# Plot one colour per fold
countries <- terra::vect(
  system.file("extdata", "world_countries.shp", package = "gbif.range")
)
countries.focus <- terra::crop(
    countries,
    terra::ext(73.0, 135.0, 18.0, 54.0)
)
terra::plot(countries.focus, col = "#bcbddc")
graphics::points(
    obs.pd[, c("decimalLongitude","decimalLatitude")],
    pch = 20,
    col = block.pd,
    cex = 1
)

}
}
\references{
Brun, P., Thuiller, W., Chauvier, Y., Pellissier, L., W\enc{ü}{u}est, R. O.,
Wang, Z., & Zimmermann, N. E. (2020). Model complexity affects species
distribution projections under climate change. Journal of Biogeography,
47(1), 130-142. \doi{10.1111/jbi.13700}
}
\seealso{
\code{\link{cv_range}}() which uses this function internally to
create cross-validation folds.
}
