% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_blocks.R
\name{make_blocks}
\alias{make_blocks}
\title{Block-wise split data into training and testing}
\usage{
make_blocks(
  nfolds = 5,
  df = data.frame(),
  nblocks = nfolds * 2,
  npoints = NA,
  pres = numeric()
)
}
\arguments{
\item{nfolds}{Numeric. Number of approximately equal-sized classes (folds)
to separate groups for block-cross validation.}

\item{df}{Object of class \code{data.frame} with \emph{n} columns
containing criteria for cluster building. Not necessary if parameter
\code{npoints} is supplied.}

\item{nblocks}{Numeric. Number of clusters (blocks) based on the number
of folds that should be built. Minimum is the same number as \code{nfolds}.
Maximum is \code{nrow(df) / 10}.}

\item{npoints}{Optional argument if \code{df} is not supplied. For how many
points should random sampling be made?}

\item{pres}{Binary vector. Optional argument. If \code{df} is supplied, this
argument can be used to save processing time. 1 stands for the points
on which CLARA is applied, and 0 stands for the points on which K-nearest
neighbors is applied relative to the 1. If \code{df} is not supplied, for
which points should random sampling be made?}
}
\value{
Object of class vector of length \code{nrow(df)} or \code{npoints},
with integers defining different folds.
}
\description{
Creates a stratum vector based on a data.frame with \emph{n} columns.
}
\details{
If \code{df} has one column, the output vector is created based on
clusters separated by quantiles. If \code{df} has two or more columns,
the output vector is created based on the 'Clustering Large Applications'
method [\code{clara()} from the \code{cluster} R package]. Also,
instead of a \code{data.frame} the argument \code{npoints} can be provided,
which create groups by random sampling. An opitimization algorithm optimizes
for equal stratum sizes (\code{gridSearch} from the \code{NMOF} R package)
}
\examples{
\dontrun{
# Downloading worldwide the observations of Panthera tigris
obs.pt <- get_gbif(
    sp_name = "Panthera tigris",
    basis = c("OBSERVATION","HUMAN_OBSERVATION",
        "MACHINE_OBSERVATION","OCCURRENCE")
)

# Create a vector of folds (n = 5) spatially blocked (n = 10)
block.pt <- make_blocks(
    nfolds = 5,
    df = obs.pt[, c("decimalLatitude","decimalLongitude")],
    nblocks = 10
)

# Plot one colour per fold
countries <- rnaturalearth::ne_countries(
    type = "countries",
    returnclass = "sv"
)
countries.focus <- terra::crop(
    countries,
    terra::ext(60,100,0,40)
)
terra::plot(countries.focus, col = "#bcbddc")
graphics::points(
    obs.pt[, c("decimalLongitude","decimalLatitude")],
    pch = 20,
    col = block.pt,
    cex = 1
)
}
}
\references{
Brun, P., Thuiller, W., Chauvier, Y., Pellissier, L., Wüest, R. O.,
Wang, Z., & Zimmermann, N. E. (2020). Model complexity affects species
distribution projections under climate change. Journal of Biogeography,
47(1), 130-142.
}
