% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ecoregion.R
\name{make_ecoregion}
\alias{make_ecoregion}
\title{Make an ecoregion map based on input environmental variables}
\usage{
make_ecoregion(
  env = NULL,
  nclass = NULL,
  path = "",
  name = "",
  raster = FALSE,
  ...
)
}
\arguments{
\item{env}{Object of class SpatRaster, RasterBrick or RasterStack of desired
resolution, crs and extent defining the study area. Used to generate a map of
clusters summarizing the environmental space of the study area.}

\item{nclass}{Numeric, How many number of environmental classes should have
the output?}

\item{path}{Character. Folder path where the output should be saved. Default
is none.}

\item{name}{Character. If 'path' is used, should include the name of the output
file (without file extension)}

\item{raster}{Logical. Whether the output should be a raster layer. Default
is FALSE.}

\item{...}{Additonnal parameters for the function clara() of the clutser R package.}
}
\value{
A TIFF or SHP file
}
\description{
This function may be used if the in-house ecoregion shapefiles are
too coarse for a given geographic region (e.g., for local studies) or
a shapefile of finer environmental details is needed. Based on several
environmental layers (e.g. climate, soil and land cover), this function
can generate a map of environmental regions containing n categories/classes.
The classes are calculated with the 'clustering large applications'
method (CLARA), which recognize patterns and relationships existing in
spatial data, and classify it in clusters.
}
\examples{
# Open
rst.path <- paste0(system.file(package = "gbif.range"),"/extdata/rst_enl.tif")
rst <- terra::rast(rst.path)
shp.path <- paste0(system.file(package = "gbif.range"),"/extdata/shp_lonlat.shp")
shp.lonlat <- terra::vect(shp.path)
rst <- terra::crop(rst, shp.lonlat)

# Apply the function by infering 50 classes of environments
my.eco <- make_ecoregion(rst,50)
terra::plot(my.eco)

# Downloading in the European Alps the observations of one plant species
obs.arcto <- get_gbif("Arctostaphylos alpinus", geo = shp.lonlat)

# Create the range map based on our custom ecoregion
range.arcto <- get_range(obs.arcto, my.eco, "EcoRegion", res = 20)

# Plot
terra::plot(shp.lonlat, col = "grey")
terra::plot(range.arcto[[2]], add = TRUE, col = rgb(0.2,1,0.2,0.5,1))
graphics::points(obs.arcto[,c("decimalLongitude","decimalLatitude")],
	pch = 20, col = "orange1", cex = 1)
}
\references{
Chauvier, Y., Zimmermann, N. E., Poggiato, G., Bystrova, D., Brun, P., &
Thuiller, W. (2021). Novel methods to correct for observer and sampling bias
in presence‐only species distribution models. Global Ecology and Biogeography,
30(11), 2312-2325.

Maechler, M., Rousseeuw, P., Struyf, A., Hubert, M., & Hornik, K. (2021). cluster:
Cluster Analysis Basics and Extensions. R package version 2.1.2 — For new features,
see the 'Changelog' file (in the package source). https://CRAN.R-project.org/package=cluster

Reynolds, A. P., Richards, G., de la Iglesia, B., & Rayward-Smith, V. J. (2006). Clustering
rules: A comparison of partitioning and hierarchical clustering algorithms. Journal of
Mathematical Modelling and Algorithms, 5(4), 475–504. doi: 10.1007/s10852-005-9022-1

Schubert, E., & Rousseeuw, P. J. (2019). Faster k-Medoids clustering: Improving the PAM,
CLARA, and CLARANS algorithms. In G. Amato, C. Gennaro, V. Oria, & M. Radovanović (Eds.),
Similarity search and applications. SISAP 2019. Lecture Notes in Computer Science (Vol.
11807, pp. 171–187). Springer.
}
