% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tiles.R
\name{make_tiles}
\alias{make_tiles}
\title{Create Tiled GBIF Geometry Queries}
\usage{
make_tiles(geo, ntiles, sext = TRUE)
}
\arguments{
\item{geo}{Spatial extent or geometry used to define the study area.
Accepted classes are \code{Extent}, \code{SpatExtent},
\code{SpatialPolygon}, \code{SpatialPolygonDataFrame}, \code{SpatVector},
and \code{sf}. If \code{NULL}, the whole globe is used.}

\item{ntiles}{Numeric approximate number of tiles to create.}

\item{sext}{Logical. Should the corresponding \code{SpatExtent} objects also
be returned?}
}
\value{
If \code{sext = TRUE}, a list containing WKT \code{POLYGON()}
strings and matching \code{SpatExtent} objects. Otherwise, a vector of WKT
\code{POLYGON()} strings.
}
\description{
Divide a study extent into a set of smaller \code{POLYGON()} query strings
that can be used with the GBIF \code{geometry} parameter.
}
\details{
The input extent is converted into a regular grid of smaller query
polygons. This is mainly intended to support tiled GBIF downloads when large
extents would otherwise return too many records in a single query.
}
\examples{
# Load the European Alps Extent
shp.path <- paste0(
	system.file(package = "gbif.range"), 
	"/extdata/shp_lonlat.shp"
)
shp.lonlat <- terra::vect(shp.path)
 
# Apply the function to divide the extent in ~20 fragments
mt <- make_tiles(geo = shp.lonlat, ntiles = 20, sext = TRUE)
}
\references{
Chauvier, Y., Thuiller, W., Brun, P., Lavergne, S., Descombes, P.,
Karger, D. N., ... & Zimmermann, N. E. (2021). Influence of climate,
soil, and land cover on plant species distribution in the European Alps.
Ecological monographs, 91(2), e01433. 10.1002/ecm.1433
}
