% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tiles.R
\name{make_tiles}
\alias{make_tiles}
\title{Create a specific number of tiles based on an extent}
\usage{
make_tiles(geo, Ntiles, meta = TRUE)
}
\description{
Not to be called directly by the user
}
\author{
Yohann Chauvier
}
