% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tiles.R
\name{make_tiles}
\alias{make_tiles}
\title{Generate Spatial tiles from a raster extent or geometry}
\usage{
make_tiles(geo, ntiles, sext = TRUE)
}
\arguments{
\item{geo}{Object of class \code{Extent}, \code{SpatExtent},
\code{SpatialPolygon}, \code{SpatialPolygonDataframe} or \code{SpaVector}
(WGS84 or planar) to define the study's area extent.
Default is \code{NULL} i.e. the whole globe.}

\item{ntiles}{Numeric. In how many tiles/fragments should geo be
divided approximately?}

\item{sext}{Logical. Should a list of SpatExtent also be returned
for each generated \code{POLYGON()}?}
}
\value{
A list of geometry arguments POLYGON() of length \code{ntiles},
and of \code{SpatExtent} if \code{sext = TRUE}).
}
\description{
Divides a given spatial extent or geometry into a specified number of tiles.
Each tile is returned as a \code{POLYGON()} geometry. The original extent
can be: (1) converted into a single \code{POLYGON()}, or (2) subdivided
into approximately \code{ntiles} regular fragments, each returned as a
\code{POLYGON()} and, optionally, as a \code{SpatExtent}. Based on a
specific extent, one or several tiles are generated. Tiles can be smaller
raster extents or geometry arguments \code{POLYGON()}. The original extent
is therefore either converted into a \code{POLYGON()} argument, or divided
into \code{ntiles} of regular fragments which are converted into
\code{POLYGON()} arguments and smaller SpatExtent.
}
\examples{
# Load the European Alps Extent
shp.path <- paste0(
	system.file(package = "gbif.range"), 
	"/extdata/shp_lonlat.shp"
)
shp.lonlat <- terra::vect(shp.path)
 
# Apply the function to divide the extent in ~20 fragments
mt <- make_tiles(geo = shp.lonlat, ntiles = 20, sext = TRUE)
}
\references{
Chauvier, Y., Thuiller, W., Brun, P., Lavergne, S., Descombes, P.,
Karger, D. N., ... & Zimmermann, N. E. (2021). Influence of climate,
soil, and land cover on plant species distribution in the European Alps.
Ecological monographs, 91(2), e01433. 10.1002/ecm.1433
}
