% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optme.R
\name{optme}
\alias{optme}
\title{Optimization function to create equal-sized strata in the 'make_blocks' function}
\usage{
optme(x, nms, grps, tot)
}
\arguments{
\item{x}{Input 1.}

\item{nms}{Input 2.}

\item{grps}{Input 3.}

\item{tot}{Input 4.}
}
\description{
Not to be called directly by the user.
}
\keyword{internal}
