% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsl_doi.R
\name{wsl_doi}
\alias{wsl_doi}
\title{Get a custom DOI for a GBIF filtered dataset}
\usage{
wsl_doi(
  wsl_gbif = NULL,
  title = NULL,
  description = NULL,
  source_url = "https://example.com/",
  user = "",
  pwd = "",
  ...
)
}
\arguments{
\item{wsl_gbif}{data.frame or list. One wsl_gbif() output or a list of several.}

\item{title}{The title for your derived dataset.}

\item{source_url}{A link to where the dataset is stored.}

\item{user}{Your GBIF username.}

\item{pwd}{Your GBIF password.}

\item{...}{Additonnal parameters for derived_dataset() in rgbif.}

\item{descritpion}{A description of the dataset.}
}
\value{
One citable DOI and its information.
}
\description{
A small user friendly wrapper of the derived_dataset() function of the
rgbif R package, compatible with one or several wsl_gbif() outputs.
}
\details{
see derived_dataset() function from the rgbif R package
}
\examples{

# Downloading worldwide the observations of Panthera tigris and Ailuropoda melanoleuca
obs.pt = wsl_gbif("Panthera tigris")
obs.am = wsl_gbif("Ailuropoda melanoleuca")

# Just an example on how to retrieve the DOI for only one wsl_gbif() output
wsl_doi(obs.pt,title="GBIF_test1",description="A small example 1",
    source_url="https://example.com/",user="",pwd="") # Use your own GBIF credentials here

# Just an example on how to retrieve the DOI for several wsl_gbif() outputs
wsl_doi(list(obs.pt,obs.am),title="GBIF_test2",description="A small example 2",
    source_url="https://example.com/",user="",pwd="") # Use your own GBIF credentials here

}
\references{
Chamberlain, S., Oldoni, D., & Waller, J. (2022). rgbif: interface to the global biodiversity
information facility API. 10.5281/zenodo.6023735
}
\seealso{
The rgbif package for additional and more general approaches to get GBIF DOI
}
